% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{int value of the state the user wishes to filter by.}

\item{year}{int value that user wished to provide the summarization for.}
}
\value{
data.frame provides a summarization of fatalities grouped by year, and month dimensions.
}
\description{
this function returns a summarization of fatalities grouped by the year and month, and accepts alist of years.
}
\examples{
\dontrun{
 fars_map_state(state.num = 1, year = 2015)
}
}
