% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{string name of file to be imported.}
}
\value{
data.frame
}
\description{
attempts to read a file and creates a data.frame from that file if it is found to exist.
}
