% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{string name of file to be imported.}
}
\value{
string of file name.
}
\description{
accepts a list of years to get car fatality data for.
}
