% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{int value that user wished to provide the summarization for.}
}
\value{
data.frame provides a summarization of fatalities grouped by year, and month dimensions.
}
\description{
this function returns a summarization of fatalities grouped by the year and month, and accepts alist of years.
}
