% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{string name of file to be imported.}
}
\value{
string of file name.
}
\description{
appends a year (int) to the file name as a string. will fail in case where a non-integer value or null is paaed to the year parameter.
}
